requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
passlib[bcrypt]
python-dotenv
slowapi
python-multipart
cloudinary
email-validator
